Core: fix the circular import issue by separating redis utils out of services
