+++ python/cloudtik/core/_private/runtime_utils.py
@@ -5,7 +5,6 @@ from typing import Dict, Any
 
 import yaml
 
-from cloudtik.core._private import services
 from cloudtik.core._private.constants import CLOUDTIK_RUNTIME_ENV_NODE_TYPE, CLOUDTIK_RUNTIME_ENV_NODE_IP, \
     CLOUDTIK_RUNTIME_ENV_SECRETS, CLOUDTIK_RUNTIME_ENV_HEAD_IP, env_bool, CLOUDTIK_DATA_DISK_MOUNT_POINT, \
     CLOUDTIK_DATA_DISK_MOUNT_NAME_PREFIX, CLOUDTIK_DEFAULT_PORT, CLOUDTIK_REDIS_DEFAULT_PASSWORD, \
@@ -16,6 +15,7 @@ from cloudtik.core._private.provider_factory import _get_node_provider
 from cloudtik.core._private.service_discovery.naming import _get_cluster_node_fqdn_of, _get_cluster_node_sdn_of, \
     get_address_type_of_hostname
 from cloudtik.core._private.service_discovery.utils import ServiceAddressType
+from cloudtik.core._private.redis_utils import create_redis_client
 from cloudtik.core._private.state.state_utils import NODE_STATE_NODE_IP, NODE_STATE_NODE_SEQ_ID
 from cloudtik.core._private.utils import load_head_cluster_config, _get_node_type_specific_runtime_config, \
     get_runtime_config_key, decode_cluster_secrets, CLOUDTIK_CLUSTER_NODES_INFO_NODE_TYPE, \
@@ -285,7 +285,7 @@ def get_cluster_redis_address():
 def get_redis_client(redis_address=None, redis_password=None):
     if not redis_address:
         redis_address, redis_password = get_cluster_redis_address()
-    return services.create_redis_client(
+    return create_redis_client(
         redis_address, redis_password)
